// src/internals/api/schema.d.ts
export interface paths {
  '/span': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    get: {
      parameters: {
        query: {
          include_attributes_ctx?: boolean;
          trace_id: string;
        };
        header?: never;
        path?: never;
        cookie?: never;
      };
      requestBody?: never;
      responses: {
        /** @description Default Response */
        200: {
          headers: {
            [name: string]: unknown;
          };
          content: {
            'application/json': {
              results: {
                attributes: {
                  ctx?: {
                    [key: string]: unknown;
                  } | null;
                  data?: {
                    [key: string]: unknown;
                  } | null;
                  target: string;
                };
                context: {
                  span_id: string;
                };
                /** Format: date-time */
                end_time: string;
                name: string;
                parent_id?: string;
                /** Format: date-time */
                start_time: string;
                /** @enum {string} */
                status_code: 'ERROR' | 'OK';
                status_message?: string;
              }[];
              total_count: number;
            };
          };
        };
        /** @description Server could not understand the request due to invalid syntax. In most cases relates with the schema validation. */
        400: {
          headers: {
            [name: string]: unknown;
          };
          content: {
            'application/json': components['schemas']['BadRequestResponse'];
          };
        };
        /** @description Unauthorized route access. */
        401: {
          headers: {
            [name: string]: unknown;
          };
          content: {
            'application/json': components['schemas']['UnauthorizedResponse'];
          };
        };
        /** @description The server encountered an unexpected condition that prevented it from fulfilling the request. */
        500: {
          headers: {
            [name: string]: unknown;
          };
          content: {
            'application/json': components['schemas']['InternalServerErrorResponse'];
          };
        };
      };
    };
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
  schemas: {
    BadRequestResponse: components['schemas']['BaseErrorResponse'] & {
      /** @enum {unknown} */
      code: 'INVALID_INPUT';
    };
    BaseErrorResponse: {
      code: string;
      data?: {
        [key: string]: unknown;
      };
      message: string;
      reason?: string;
    };
    InternalServerErrorResponse: components['schemas']['BaseErrorResponse'] & {
      /** @enum {unknown} */
      code: 'INTERNAL_SERVER_ERROR';
    };
    NotFoundResponse: components['schemas']['BaseErrorResponse'] & {
      /** @enum {unknown} */
      code: 'NOT_FOUND';
    };
    UnauthorizedResponse: components['schemas']['BaseErrorResponse'] & {
      /** @enum {unknown} */
      code: 'AUTH_ERROR';
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}